scope sortSaleList fn to Sale view, and reorganize mock fixture into /testing
