+++ frontend/src/utils/testing/fixtures/mockData.ts
@@ -1,0 +1,315 @@
+import { ScryfallCard } from '../../ScryfallCard';
+
+const mockData: ScryfallCard[] = [
+    {
+        id: '7f78a570-d776-42f2-a609-6da0156c8de7',
+        tcgplayer_id: 198863,
+        name: 'Curious Pair // Treats to Share',
+        lang: 'en',
+        image_uris: {
+            normal: 'https://c1.scryfall.com/file/scryfall-cards/normal/front/7/f/7f78a570-d776-42f2-a609-6da0156c8de7.jpg?1572490513',
+        },
+        type_line: 'Creature — Human Peasant // Sorcery — Adventure',
+        colors: ['G'],
+        color_identity: ['G'],
+        keywords: [],
+        card_faces: [],
+        finishes: ['nonfoil', 'foil'],
+        set: 'eld',
+        set_name: 'Throne of Eldraine',
+        rarity: 'common',
+        border_color: 'black',
+        cardImage: '',
+        display_name: '',
+        frame_effects: [],
+        printed_name: '',
+        promo_types: [],
+        qoh: {},
+    },
+    {
+        id: '20f79f1a-690f-4ec1-86b8-49aa7ae8ff08',
+        tcgplayer_id: 222585,
+        name: 'Dimir Keyrune',
+        lang: 'en',
+        image_uris: {
+            normal: 'https://c1.scryfall.com/file/scryfall-cards/normal/front/2/0/20f79f1a-690f-4ec1-86b8-49aa7ae8ff08.jpg?1604195572',
+        },
+        type_line: 'Artifact',
+        colors: [],
+        color_identity: ['B', 'U'],
+        keywords: [],
+        finishes: ['nonfoil', 'foil'],
+        set: 'znc',
+        set_name: 'Zendikar Rising Commander',
+        rarity: 'uncommon',
+        border_color: 'black',
+        qoh: {},
+        promo_types: [],
+        printed_name: '',
+        frame_effects: [],
+        display_name: '',
+        cardImage: '',
+        card_faces: [],
+    },
+    {
+        id: '0e0d989d-7186-40dc-bdfe-cfbb77656bc8',
+        tcgplayer_id: 179448,
+        name: 'Emrakul, the Aeons Torn',
+        lang: 'en',
+        image_uris: {
+            normal: 'https://img.scryfall.com/cards/normal/front/0/e/0e0d989d-7186-40dc-bdfe-cfbb77656bc8.jpg?1559959218',
+        },
+        type_line: 'Legendary Creature — Eldrazi',
+        colors: [],
+        color_identity: [],
+        finishes: ['foil', 'nonfoil'],
+        set: 'uma',
+        set_name: 'Ultimate Masters',
+        rarity: 'mythic',
+        border_color: 'black',
+        frame_effects: ['legendary'],
+        qoh: {},
+        cardImage: '',
+        card_faces: [],
+        display_name: '',
+        keywords: [],
+        printed_name: '',
+        promo_types: [],
+    },
+    {
+        color_identity: ['G'],
+        colors: [],
+        id: '25b0b816-0583-44aa-9dc5-f3ff48993a51',
+        image_uris: {
+            normal: 'https://img.scryfall.com/cards/normal/front/2/5/25b0b816-0583-44aa-9dc5-f3ff48993a51.jpg?1562902898',
+        },
+        lang: 'en',
+        name: "Gaea's Cradle",
+        finishes: ['nonfoil'],
+        qoh: {
+            NONFOIL_NM: 3,
+        },
+        rarity: 'rare',
+        set: 'usg',
+        set_name: "Urza's Saga",
+        tcgplayer_id: 6883,
+        type_line: 'Legendary Land',
+        border_color: '',
+        cardImage: '',
+        card_faces: [],
+        display_name: '',
+        frame_effects: [],
+        keywords: [],
+        printed_name: '',
+        promo_types: [],
+    },
+    {
+        id: 'fcac0cff-efa6-4068-a69c-67ea4e7acea0',
+        tcgplayer_id: 179507,
+        name: 'Sigarda, Host of Herons',
+        lang: 'en',
+        image_uris: {
+            normal: 'https://img.scryfall.com/cards/normal/front/f/c/fcac0cff-efa6-4068-a69c-67ea4e7acea0.jpg?1559959280',
+        },
+        type_line: 'Legendary Creature — Angel',
+        colors: ['G', 'W'],
+        color_identity: ['G', 'W'],
+        finishes: ['foil', 'nonfoil'],
+        set: 'uma',
+        set_name: 'Ultimate Masters',
+        rarity: 'mythic',
+        border_color: 'black',
+        frame_effects: ['legendary'],
+        qoh: {
+            NONFOIL_NM: 3,
+        },
+        cardImage: '',
+        card_faces: [],
+        display_name: '',
+        keywords: [],
+        printed_name: '',
+        promo_types: [],
+    },
+    {
+        id: 'd28056c7-c58d-4986-a45c-c9e55aed23a1',
+        tcgplayer_id: 155769,
+        name: 'Colossal Dreadmaw',
+        lang: 'en',
+        image_uris: {
+            normal: 'https://img.scryfall.com/cards/normal/front/d/2/d28056c7-c58d-4986-a45c-c9e55aed23a1.jpg?1555040601',
+        },
+        type_line: 'Creature — Dinosaur',
+        colors: ['G'],
+        color_identity: ['G'],
+        finishes: ['foil', 'nonfoil'],
+        set: 'rix',
+        set_name: 'Rivals of Ixalan',
+        rarity: 'common',
+        border_color: 'black',
+        qoh: {
+            NONFOIL_NM: 5,
+        },
+        cardImage: '',
+        card_faces: [],
+        display_name: '',
+        frame_effects: [],
+        keywords: [],
+        printed_name: '',
+        promo_types: [],
+    },
+    {
+        border_color: 'black',
+        color_identity: ['R'],
+        colors: ['R'],
+        finishes: ['nonfoil'],
+        id: '9052369f-840f-438e-b86d-e2f8d6339585',
+        image_uris: {
+            normal: 'https://img.scryfall.com/cards/normal/front/9/0/9052369f-840f-438e-b86d-e2f8d6339585.jpg?1559591625',
+        },
+        lang: 'en',
+        name: 'Wheel of Fortune',
+        qoh: {
+            NONFOIL_NM: 4,
+        },
+        rarity: 'rare',
+        set: 'leb',
+        set_name: 'Limited Edition Beta',
+        tcgplayer_id: 8962,
+        type_line: 'Sorcery',
+        cardImage: '',
+        card_faces: [],
+        display_name: '',
+        frame_effects: [],
+        keywords: [],
+        printed_name: '',
+        promo_types: [],
+    },
+    {
+        id: 'c89c6895-b0f8-444a-9c89-c6b4fd027b3e',
+        tcgplayer_id: 145298,
+        name: 'Necropotence',
+        lang: 'en',
+        image_uris: {
+            normal: 'https://img.scryfall.com/cards/normal/front/c/8/c89c6895-b0f8-444a-9c89-c6b4fd027b3e.jpg?1562853736',
+        },
+        type_line: 'Enchantment',
+        colors: ['B'],
+        color_identity: ['B'],
+        finishes: ['foil', 'nonfoil'],
+        set: 'ima',
+        set_name: 'Iconic Masters',
+        rarity: 'mythic',
+        border_color: 'black',
+        qoh: {
+            NONFOIL_NM: 2,
+        },
+        cardImage: '',
+        card_faces: [],
+        display_name: '',
+        frame_effects: [],
+        keywords: [],
+        printed_name: '',
+        promo_types: [],
+    },
+    {
+        id: '9d75faf7-fc27-4fc2-9e80-e35232c42542',
+        tcgplayer_id: 190980,
+        name: 'Bazaar Trademage',
+        lang: 'en',
+        image_uris: {
+            normal: 'https://img.scryfall.com/cards/normal/front/9/d/9d75faf7-fc27-4fc2-9e80-e35232c42542.jpg?1562201307',
+        },
+        type_line: 'Creature — Human Wizard',
+        colors: ['U'],
+        color_identity: ['U'],
+        finishes: ['foil', 'nonfoil'],
+        set: 'mh1',
+        set_name: 'Modern Horizons',
+        rarity: 'rare',
+        border_color: 'black',
+        qoh: {
+            FOIL_NM: 1,
+            NONFOIL_NM: 1,
+        },
+        cardImage: '',
+        card_faces: [],
+        display_name: '',
+        frame_effects: [],
+        keywords: [],
+        printed_name: '',
+        promo_types: [],
+    },
+    {
+        border_color: 'black',
+        image_uris: { normal: '' },
+        colors: ['R', 'W'],
+        card_faces: [
+            {
+                type_line: 'Legendary Creature — Angel',
+
+                colors: ['W'],
+
+                image_uris: {
+                    normal: 'https://img.scryfall.com/cards/normal/front/a/e/ae155ee2-008f-4dc6-82bf-476be7baa224.jpg?1576383667',
+                },
+                color_identity: ['R', 'W'],
+            },
+            {
+                type_line: 'Legendary Creature — Angel',
+
+                colors: ['R'],
+                color_identity: ['R', 'W'],
+
+                image_uris: {
+                    normal: 'https://img.scryfall.com/cards/normal/back/a/e/ae155ee2-008f-4dc6-82bf-476be7baa224.jpg?1576383667',
+                },
+            },
+        ],
+        color_identity: ['R', 'W'],
+        frame_effects: ['sunmoondfc'],
+        id: 'ae155ee2-008f-4dc6-82bf-476be7baa224',
+        lang: 'en',
+        name: 'Archangel Avacyn // Avacyn, the Purifier',
+        finishes: ['foil', 'nonfoil'],
+        qoh: {
+            NONFOIL_NM: 1,
+        },
+        rarity: 'mythic',
+        set: 'soi',
+        set_name: 'Shadows over Innistrad',
+        tcgplayer_id: 114848,
+        type_line: 'Legendary Creature — Angel // Legendary Creature — Angel',
+        cardImage: '',
+        display_name: '',
+        keywords: [],
+        printed_name: '',
+        promo_types: [],
+    },
+    {
+        id: '0020a124-ba76-4d40-84e9-9803268d9f16',
+        tcgplayer_id: 110837,
+        name: 'World Breaker',
+        lang: 'en',
+        image_uris: {
+            normal: 'https://c1.scryfall.com/file/scryfall-cards/normal/front/0/0/0020a124-ba76-4d40-84e9-9803268d9f16.jpg?1562895014',
+        },
+        type_line: 'Creature — Eldrazi',
+        colors: [],
+        color_identity: ['G'],
+        keywords: ['Reach', 'Devoid'],
+        finishes: ['foil', 'nonfoil'],
+        set: 'ogw',
+        set_name: 'Oath of the Gatewatch',
+        rarity: 'mythic',
+        border_color: 'black',
+        frame_effects: ['devoid'],
+        cardImage: '',
+        card_faces: [],
+        display_name: '',
+        printed_name: '',
+        promo_types: [],
+        qoh: {},
+    },
+];
+
+export default mockData;
